+++ examples/area_gallery.rb
@@ -27,7 +27,7 @@ window('Area Gallery', 400, 400) {
         }
 
         fill r: 202, g: 102, b: 104, a: 0.5
-        stroke thickness: 1, r: 0, g: 0, b: 0
+        stroke r: 0, g: 0, b: 0
       }
       path { # declarative stable path
         figure(0, 0) {